Initial server-side anti-cheat code
Doesn't detect known cheat mods yet, but it's a start.

Very basic in that it only checks modIDs, but the cool thing about this is that it still works when a client joins without the itlt mod. Checking hashes will require client co-operation.

Of course, all of this mod's anti-cheat measures can be bypassed with sufficient know-how (as with any anti-cheat - especially client-assisted ones) but the point of this is to compliment existing server-side-only anti-cheat systems rather than replace them.