Improve memory calc, simplify display to user
The memory calc now factors in the metaspace, making it more accurately reflect Xmx, be it an implied or explicit JVM arg. It is also rounded to the nearest tenth (e.g. 1.0, 1.1, 1.2...)

Memory display to the user is now simplified when the representation of the value is the same. For example, instead of saying "You need 2.0GB of RAM but you only have 1.0GB", it will now say "You need 2GB of RAM but you only have 1GB".